fix for #39 MD5 hash bug, UINT4 is now an uint32_t, which should work also correct on 64bit systems

+++ src/generator/md5.hpp
@@ -39,8 +39,10 @@
  **********************************************************************
  */
 #include <string>
+#include <stdint.h>
+
 /* typedef a 32 bit type */
-typedef unsigned long int UINT4;
+typedef uint32_t UINT4;
 /* Data structure for MD5 (Message Digest) computation */
 typedef struct {
   UINT4 i[2];                   /* number of _bits_ handled mod 2^64 */
